feat(nx-dev): add nx-blog sitemap to root sitemap index
## Current Behavior

The root sitemap.xml index references only /sitemap-1.xml (Framer via
edge-function proxy) and /docs/sitemap-index.xml (astro-docs). Blog
posts published by nx-blog (BLOG_URL) are not advertised to crawlers
through the root nx.dev sitemap.

## Expected Behavior

The root sitemap index additionally references /sitemap-2.xml, which is
served by a consolidated `additional-sitemaps.ts` edge function that
proxies the per-source sitemaps:

  /sitemap-1.xml -> <NEXT_PUBLIC_FRAMER_URL>/sitemap.xml
  /sitemap-2.xml -> <BLOG_URL>/blog/sitemap.xml

URLs in the upstream XML are rewritten from the source origin to
nx.dev. The previous per-source edge functions (framer-sitemap.ts,
blog-sitemap.ts) are replaced by the single additional-sitemaps.ts
to match the "additionalSitemaps" language used in the Next.js sitemap
config.

## Related Issue(s)

Fixes DOC-486
