+++ nx-dev/nx-dev/DEPLOYMENT.md
@@ -27,10 +27,10 @@ When a request hits nx.dev, it goes through the following stages in order:
 
 Edge functions run first, before any other routing. Located at `/netlify/edge-functions/` in the repository root (see [Why repo root?](#why-are-edge-functions-at-the-repo-root)).
 
-| Edge Function            | Path Pattern     | Purpose                                     |
-| ------------------------ | ---------------- | ------------------------------------------- |
-| `rewrite-framer-urls.ts` | `/*`             | Proxies marketing pages to Framer           |
-| `framer-sitemap.ts`      | `/sitemap-1.xml` | Proxies Framer's sitemap with URL rewriting |
+| Edge Function            | Path Pattern                       | Purpose                                                    |
+| ------------------------ | ---------------------------------- | ---------------------------------------------------------- |
+| `rewrite-framer-urls.ts` | `/*`                               | Proxies marketing pages to Framer                          |
+| `additional-sitemaps.ts` | `/sitemap-1.xml`, `/sitemap-2.xml` | Proxies Framer's and nx-blog's sitemaps with URL rewriting |
 
 ### 2. Netlify Redirects (`_redirects`)
 
@@ -134,7 +134,7 @@ netlify/
 └── edge-functions/
     ├── README.md
     ├── rewrite-framer-urls.ts   # Main Framer proxy
-    └── framer-sitemap.ts        # Framer sitemap proxy
+    └── additional-sitemaps.ts   # Framer + nx-blog sitemap proxy
 ```
 
 #### Why are edge functions at the repo root?
@@ -205,6 +205,7 @@ included_files = [
 | `NEXT_PUBLIC_FRAMER_URL` | Framer site URL for marketing pages           | `https://ready-knowledge-238309.framer.app`     |
 | `NEXT_PUBLIC_ASTRO_URL`  | Astro docs site URL                           | `https://master--nx-docs.netlify.app`           |
 | `NEXT_PUBLIC_BANNER_URL` | Framer CMS URL for banner data                | `https://your-site.framer.app/api/banners/main` |
+| `BLOG_URL`               | nx-blog site URL for blog/changelog pages     | `https://blog.nx.app`                           |
 | `NX_DEV_URL`             | Canonical site URL for sitemap                | `https://nx.dev`                                |
 | `NEXT_PUBLIC_NO_INDEX`   | Set to `true` to add noindex robots directive | `true`                                          |
 
@@ -213,10 +214,11 @@ included_files = [
 The sitemap is composed of multiple sources:
 
 1. **Next.js sitemap** (`/sitemap.xml`) - Generated by `next-sitemap`
-2. **Framer sitemap** (`/sitemap-1.xml`) - Proxied via `framer-sitemap.ts` edge function
-3. **Astro sitemap** (`/docs/sitemap-index.xml`) - Served by Astro
+2. **Framer sitemap** (`/sitemap-1.xml`) - Proxied via `additional-sitemaps.ts` edge function
+3. **nx-blog sitemap** (`/sitemap-2.xml`) - Proxied via `additional-sitemaps.ts` edge function
+4. **Astro sitemap** (`/docs/sitemap-index.xml`) - Served by Astro
 
-The sitemap index references these via `additionalSitemaps` in `next-sitemap.config.js`.
+The sitemap index references these via `scripts/patch-sitemap-index.mjs`, which runs after `next-sitemap` generates the index.
 
 ## Common Tasks
